updating and deleting ToDo by user id

+++ routers/todos.py
@@ -68,10 +68,16 @@ async def create_todo(user: user_dependency,
 
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency,
+async def update_todo(user: user_dependency,
+                      db: db_dependency,
                       todo_request: ToDoRequest,
                       todo_id: int = Path(gt=0)):
-    todo_model = db.query(ToDos).filter(ToDos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+
+    todo_model = db.query(ToDos).filter(ToDos.id == todo_id)\
+                                .filter(ToDos.owner_id == user.get('id'))\
+                                .first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='ToDo not found.')
 
@@ -85,8 +91,15 @@ async def update_todo(db: db_dependency,
 
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(ToDos).filter(ToDos.id == todo_id).first()
+async def delete_todo(db: db_dependency,
+                      user: user_dependency,
+                      todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+
+    todo_model = db.query(ToDos).filter(ToDos.id == todo_id)\
+                                .filter(ToDos.owner_id == user.get('id'))\
+                                .first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='ToDo not found.')
     db.query(ToDos).filter(ToDos.id == todo_id).delete()
